nginx-auth/python-auth/sources/requirements.txt
python-ldap==3.3.1
PyYAML==6.0
